+++ test/llvm-flake-parts/flake.nix
@@ -49,7 +49,7 @@
             (
             set -x
             ${config.conan.outputs.devShell.shellHook}
-            conan create ${config.conan.info.configRoot} -tf="" --build=missing 2>&1 | grep "example/0.0.1"
+            conan create ${config.conan.info.configRoot} -tf="" --build=missing 2>&1 | grep -F "example/0.0.1"
             touch $out
             )
           '';