+++ test/standalone-submodule-with/flake.nix
@@ -56,7 +56,7 @@
               (
               set -x
               ${conanModuleConfig.outputs.devShell.shellHook}
-              conan create ${conanModuleConfig.info.configRoot} -tf="" --build=missing 2>&1 | grep "example/0.0.1"
+              conan create ${conanModuleConfig.info.configRoot} -tf="" --build=missing 2>&1 | grep -F "example/0.0.1"
               touch $out
               )
             '';