+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Excluded `./examples/devenv-module/devenv.nix` from dev treefmt settings.
 - Fixed handling of root configuration path when setting local recipe index repos.
 - Set `info.configRoot` option.
+- Fixed tests to use `grep -F`.
 
 ### Improvements
 
@@ -24,11 +25,14 @@
 - Added CMake by default to `devShell.tools`.
 - Added a `defaults.profiles.platformToolRequires` option and set a `cmake` attribute by default depending whether CMake is a required tool or not.
 - `conan.stdenv.cc` added as a default tool.
+- Added `final.profiles.platformToolRequires` option.
 
 ### Breaking Changes
 
 - Removed default `info` output command.
 - Removed automatic mapping of `config.conan.outputs.packages`.
+- Removed `platformToolRequires` option (`profiles.platformToolRequires` must be used instead).
+- Removed the `buildEnv`, `runEnv` and `conf` options (keep only those in the `profiles` namespace).
 
 ## 0.3.1 (May 29, 2026)